# openapi.yaml
openapi: 3.0.3
info:
  title: Varius API
  version: 0.1.0
  description: "OpenAPI specification for the vx-api-server (summary of implemented endpoints)"
servers:
  - url: https://api.varius.technology
    description: Production server
paths:
...
  /projects:
    post:
      summary: Create a new project (new or legacy payload supported)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CreateProjectPayload'
                - $ref: '#/components/schemas/LegacyCreateProjectPayload'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Validation error
        '409':
          description: Project ID already exists
    get:
      summary: List projects (pagination) or fetch single project by projectId query
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
          description: Optional project UUID or user-specified project_id to fetch a single project
        - name: page
          in: query
          schema:
            type: integer
        - name: perPage
          in: query
          schema:
            type: integer
        - name: ownerId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated projects or a single project when projectId is provided
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProjectResponse'
                      total:
                        type: integer
                      page:
                        type: integer
                      perPage:
                        type: integer
                  - $ref: '#/components/schemas/ProjectResponse'
  /projects/{projectId}:
    get:
      summary: Get single project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
    patch:
      summary: Update project (partial)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartialProjectUpdate'
      responses:
        '200':
          description: Updated project
  /projects/{projectId}/todos/{todoId}:
    patch:
      summary: Update a todo
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: todoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                done:
                  type: boolean
                order:
                  type: integer
      responses:
        '200':
          description: Updated todo
    delete:
      summary: Delete a todo
      responses:
        '204':
          description: Deleted
components:
  schemas:
...
    CreateProjectPayload:
      type: object
      required:
        - name
        - projectId
        - creator
        - currencies
      properties:
        name:
          type: string
        projectId:
          type: string
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        nodeEndpoint:
          type: string
          nullable: true
        creator:
          oneOf:
            - type: string
            - type: integer
        currencies:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
          nullable: true
    LegacyCreateProjectPayload:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
        ownerId:
          type: string
          nullable: true
        todos:
          type: array
          items:
            type: object
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        projectId:
          type: string
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        nodeEndpoint:
          type: string
          nullable: true
        creator:
          type: string
        currencies:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
          nullable: true
        todos:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
        title:
          type: string
          nullable: true
        ownerId:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
    Todo:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        order:
          type: integer
        done:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    PartialProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        nodeEndpoint:
          type: string
          nullable: true
        creator:
          oneOf:
            - type: string
            - type: integer
        currencies:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
          nullable: true
        title:
          type: string
        ownerId:
          type: string
        dueDate:
          type: string
          format: date
...